Remove suppliers
