+++ src/test/java/org/opentripplanner/graph_builder/module/islandpruning/AdaptivePruningTest.java
@@ -9,11 +9,8 @@ import org.opentripplanner.ConstantsForTests;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
 import org.opentripplanner.graph_builder.module.osm.OpenStreetMapModule;
 import org.opentripplanner.graph_builder.module.osm.OpenStreetMapModuleBuilder;
-import org.opentripplanner.graph_builder.services.osm.CustomNamer;
 import org.opentripplanner.openstreetmap.OpenStreetMapProvider;
-import org.opentripplanner.openstreetmap.model.OSMWithTags;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.street.model.edge.StreetEdge;
 import org.opentripplanner.transit.model.framework.Deduplicator;
 import org.opentripplanner.transit.service.StopModel;
 import org.opentripplanner.transit.service.TransitModel;
@@ -77,26 +74,9 @@ public class AdaptivePruningTest {
       // Add street data from OSM
       File osmFile = new File(osmPath);
       OpenStreetMapProvider osmProvider = new OpenStreetMapProvider(osmFile, true);
-      var customNamer = new CustomNamer() {
-        @Override
-        public String name(OSMWithTags way, String defaultName) {
-          return String.valueOf(way.getId());
-        }
-
-        @Override
-        public void nameWithEdge(OSMWithTags way, StreetEdge edge) {}
-
-        @Override
-        public void postprocess(Graph graph) {}
-
-        @Override
-        public void configure() {}
-      };
       OpenStreetMapModule osmModule = OpenStreetMapModuleBuilder
         .of(osmProvider, graph)
-        .withCustomNamer(customNamer)
-        .withAreaVisibility(false)
-        .withStaticParkAndRide(false)
+        .withCustomNamer(new TestCustomNamer())
         .build();
 
       osmModule.buildGraph();
