+++ src/test/java/org/opentripplanner/graph_builder/module/osm/TriangleInequalityTest.java
@@ -58,7 +58,6 @@ public class TriangleInequalityTest {
     OpenStreetMapModule osmModule = OpenStreetMapModuleBuilder
       .of(provider, graph)
       .withAreaVisibility(true)
-      .withStaticParkAndRide(false)
       .build();
     osmModule.buildGraph();
   }
